requirements.txt
aiohttp==3.10.11
aiosqlite==0.22.1
alembic==1.18.1
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apprise==1.9.6
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.0
cssselect==1.3.0
cycler==0.12.1
DataRecorder==3.6.2
DownloadKit==2.0.7
DrissionPage==4.1.1.2
duckdb==1.1.3
email-validator>=2.0.0
et_xmlfile==2.0.0
fastapi==0.128.0
filelock==3.20.1
fonttools==4.61.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.1.3
oauthlib==3.3.1
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pefile==2024.8.26
pillow==12.0.0
platformdirs==4.5.1
prometheus-client==0.24.1
numba==0.61.0  # JIT acceleration for hamming distance
psutil==7.2.1
pyaes==1.6.1
pyarrow==19.0.0
pyasn1==0.6.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.3.1
pyperclip==1.11.0
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytokens==0.3.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
qrcode==8.2
QtAwesome==1.4.0
QtPy==2.4.3
rapidfuzz==3.12.1  # Fast string similarity
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
shiboken6==6.10.1
six==1.17.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.50.0
structlog==25.5.0
Telethon==1.42.0
tkinterdnd2==0.4.3
tldextract==5.3.1
ttkbootstrap==1.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
websockets>=11.0.3
uvloop==0.17.0
websocket-client==1.9.0
Werkzeug==3.1.4
xxhash==3.5.0  # Ultra-fast hashing
jieba==0.42.1  # Chinese word segmentation for hotwords
aiofiles==24.1.0  # Async file I/O for hotword repository
